% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisGenerator.R
\name{betaToGamma}
\alias{betaToGamma}
\title{Turn a set of coefficents in the monomial space into the corresponding
set of coefficents in the orthonormal space}
\usage{
betaToGamma(beta, R_inv)
}
\arguments{
\item{beta}{Vector of polynomial coefficients in the monomial space}

\item{R_inv}{Requisite inverse matrix}
}
\value{
Vector of the same length as beta containing the corresponding
coefficients in the orthonormal space
}
\description{
Turn a set of coefficents in the monomial space into the corresponding
set of coefficents in the orthonormal space
}
