% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revJumpHelpers.R
\name{calcAcceptProb}
\alias{calcAcceptProb}
\title{Calculate the acceptance probability for a preposal, whilst enforcing
monotoncity.}
\usage{
calcAcceptProb(gamma_prop, gamma_curr, d_prop, d_curr, d_min, d_max, var_prop,
  var_curr, Q_full, full_inits, y_vec, innov_sigma, dim_prior_vec,
  innov_sd_var = innov_sd_var, R_inv_full)
}
\arguments{
\item{gamma_prop}{Vector containing proposal for Gamma}

\item{gamma_curr}{Vector containing the current value for Gamma}

\item{d_prop}{proposed polynomial degree}

\item{d_curr}{current polynomial degree}

\item{d_min}{minimum allowable polynomial degree}

\item{d_max}{maximum allowable polynomial degree}

\item{var_prop}{proposed variance}

\item{var_curr}{current variance}

\item{Q_full}{Full size (n_obs * (d_max + 1)) matrix for Q}

\item{full_inits}{full set of initial gamma values from the MonoPoly::monpol
fit}

\item{y_vec}{Vector of y observations}

\item{innov_sigma}{innovation variance that pertains to the gamma proposal}

\item{dim_prior_vec}{Vector containing prior values for our dimension
parameter}

\item{innov_sd_var}{innovation variance pertaining to the variance proposal}

\item{R_inv_full}{Full size (d_max + 1) * (d_max + 1) basis inverse matrix}
}
\value{
An acceptance probability
}
\description{
Calculate the acceptance probability for a preposal, whilst enforcing
monotoncity.
}
