% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroHastingHelpers.R
\name{calcPostRatio}
\alias{calcPostRatio}
\title{Calculate the section of the acceptance probabilty pertaining to the
posterior distribution of our parameters}
\usage{
calcPostRatio(gamma_prop, gamma_curr, var_prop, var_curr, Q_full, d_prop,
  d_curr, y_vec, dim_prior_vec, R_inv_full)
}
\arguments{
\item{gamma_prop}{Vector containing proposal for Gamma}

\item{gamma_curr}{Vector containing the current value for Gamma}

\item{var_prop}{Proposed value for the variance}

\item{var_curr}{Current value for the variance}

\item{Q_full}{Full size (n_obs * (d_max + 1)) matrix for Q}

\item{d_prop}{Proposed polynomial degree}

\item{d_curr}{Current polynomial degree}

\item{y_vec}{Vector of y observations}

\item{dim_prior_vec}{Vector containing prior values for our dimension
parameter}

\item{R_inv_full}{Full size (d_max + 1) * (d_max + 1) basis inverse matrix}
}
\value{
A value for the posterior acceptance ratio
}
\description{
Calculate the section of the acceptance probabilty pertaining to the
posterior distribution of our parameters
}
