% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroHastingHelpers.R
\name{calcPropRatio}
\alias{calcPropRatio}
\title{Calculate the part of the acceptance probability pertaining to the proposal
distributions. This function will make absolutely no sense unless looked at
after one has consulted the maths.}
\usage{
calcPropRatio(gamma_k_prime_prop, k_prime_init, innov_sigma, d_prop, d_curr,
  d_min, d_max, var_prop, var_curr, innov_sd_var)
}
\arguments{
\item{gamma_k_prime_prop}{proposed value of gamma from the additional
dimension}

\item{k_prime_init}{mean value for the independent proposal we are using
for the additional regression parameter}

\item{innov_sigma}{innovation variance that pertains to the gamma proposal}

\item{d_prop}{proposed polynomial degree}

\item{d_curr}{current polynomial degree}

\item{d_min}{minimum allowable polynomial degree}

\item{d_max}{maximum allowable polynomial degree}

\item{var_prop}{proposed variance}

\item{var_curr}{current variance}

\item{innov_sd_var}{innovation variance pertaining to the variance proposal}
}
\value{
A value for the proposal acceptance ratio
}
\description{
Calculate the part of the acceptance probability pertaining to the proposal
distributions. This function will make absolutely no sense unless looked at
after one has consulted the maths.
}
