% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revJumpHelpers.R
\name{dimProposer}
\alias{dimProposer}
\title{Proposes new polynomial degree / dimension for parameter space}
\usage{
dimProposer(d_current, d_min, d_max)
}
\arguments{
\item{d_current}{Current polynomial degree}

\item{d_min}{minimum allowable polynomial degree}

\item{d_max}{maximum allowable polynomial degree}
}
\value{
Proposed polynomial degree
}
\description{
Proposes new polynomial degree / dimension for parameter space
}
