% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisGenerator.R
\name{gammaToBeta}
\alias{gammaToBeta}
\title{Turn a set of coefficents in the othronormal space into the corresponding
set of coefficents in the monomial space}
\usage{
gammaToBeta(gamma, R_inv)
}
\arguments{
\item{gamma}{Vector of polynomial coefficients in the orthonormal space}

\item{R_inv}{Requisite inverse matrix}
}
\value{
Vector of the same length as gamma containing the corresponding
coefficients in the monomial space
}
\description{
Turn a set of coefficents in the othronormal space into the corresponding
set of coefficents in the monomial space
}
