% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisGenerator.R
\name{genAllMatrices}
\alias{genAllMatrices}
\title{Generate the orthonormal polynomial basis matrix Q and the requisite inverse
matrix R_inv.}
\usage{
genAllMatrices(x, d)
}
\arguments{
\item{x}{A vector of x values}

\item{d}{Degree of polynomial basis}
}
\value{
A named list consisting of Q and R_inv(erse)
}
\description{
Generate the orthonormal polynomial basis matrix Q and the requisite inverse
matrix R_inv.
}
\examples{
genAllMatrices(runif(10, 0 ,1), 6)
}
