% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroHastingHelpers.R
\name{genDimPrior}
\alias{genDimPrior}
\title{generates a discrete distribution over the allowable degree space, uses
the binomial distribution to adjust the distribution}
\usage{
genDimPrior(d_min, d_max, p)
}
\arguments{
\item{d_min}{minimum allowable polynomial degree}

\item{d_max}{maximum allowable polynomial degree}

\item{p}{Probability passed to \link{dbinom}}
}
\value{
vector containing prior values
}
\description{
generates a discrete distribution over the allowable degree space, uses
the binomial distribution to adjust the distribution
}
