% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revJumpHelpers.R
\name{genGammaProp}
\alias{genGammaProp}
\title{Generates a proposal for Gamma, whilst considering proposed dimension}
\usage{
genGammaProp(current_gamma, d_prop, d_curr, innov_sigma, full_inits)
}
\arguments{
\item{current_gamma}{Vector containing the current value for Gamma (really
should rename this to gamma_curr like I have everywhere else)}

\item{d_prop}{proposed polynomial degree}

\item{d_curr}{current polynomial degree}

\item{innov_sigma}{innovation variance that pertains to the gamma proposal}

\item{full_inits}{full set of initial gamma values from the MonoPoly::monpol
fit}

\item{R_inv_full}{full size (d_max + 1) * (d_max + 1) inverse matrix}
}
\value{
Proposal for gamma
}
\description{
Generates a proposal for Gamma, whilst considering proposed dimension
}
