% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroHastingHelpers.R
\name{genSingleGammaProp}
\alias{genSingleGammaProp}
\title{Generate a proposal for our coefficients in the orthonormal space.}
\usage{
genSingleGammaProp(gamma_old, innov_sigma)
}
\arguments{
\item{gamma_old}{Vector containing the current value for gamma in the Markov
chain}

\item{innov_sigma}{Innovation variance associated with the proposal
distribution for sigma}
}
\value{
A proposed realisation for gamma
}
\description{
Generate a proposal for our coefficients in the orthonormal space.
}
