% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroHastingHelpers.R
\name{genVarProp}
\alias{genVarProp}
\title{Generates a proposal for our variance parameter}
\usage{
genVarProp(var_prev, innov_sd_var)
}
\arguments{
\item{var_prev}{Current value of variance}

\item{innov_sd_var}{Variance associated with the innovation distribution for
our variance parameter}
}
\value{
A proposed realisation for the variance
}
\description{
Generates a proposal for our variance parameter
}
