% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plotFit}
\alias{plotFit}
\title{Plot of fitted curve, with credible interval. Either uses the posterior
mode degree, or degree closest to posetrior mean.}
\usage{
plotFit(obj, mode = "mean", degree = NA, ...)
}
\arguments{
\item{obj}{object of type}

\item{mode}{which degree to plot, "mean" for rounded mean, "mode" for
posterior mode}

\item{degree}{(optional) Specify a specific degree to plot}

\item{...}{}
}
\value{
obj
}
\description{
Plot of fitted curve, with credible interval. Either uses the posterior
mode degree, or degree closest to posetrior mean.
}
