% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{returnSamplesOfLengthN}
\alias{returnSamplesOfLengthN}
\title{Return all samples in list, of length N}
\usage{
returnSamplesOfLengthN(sample_list, N)
}
\arguments{
\item{sample_list}{list containing vectors of various length}

\item{N}{Length of vectors you wish to extract}
}
\value{
Matrix with N columns and an unknown number of rows
}
\description{
Return all samples in list, of length N
}
