% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitter.R
\name{rjmonopoly}
\alias{rjmonopoly}
\title{Reversible jump sampler for monotonic polynomials}
\usage{
rjmonopoly(x, y, d_min = 2, d_max = 10, iter = 50000, prior_option = NA,
  prior_prob = 0.5, starting_var_val = 0.001, control = list(innov_sd_beta
  = 0.025, innov_sd_var = 0.025), progress = TRUE)
}
\arguments{
\item{x}{vector of x values}

\item{y}{vector of y values}

\item{d_min}{minimum allowable polynomial degree}

\item{d_max}{maximum allowable polynomial degree}

\item{iter}{number of iterations}

\item{prior_option}{if set to string "flat", then will use a flat prior over
the allowable parameter space.}

\item{prior_prob}{probability for binomial prior}

\item{starting_var_val}{starting value for variance}

\item{control}{(optional) list for innovation variances}

\item{progress}{turn the progress bar on / off, TRUE is default, and shows
progress bar.}
}
\value{
object of type rjmonopol_fit containing lists of samples for all
regression parameters
}
\description{
Other notes
}
\details{
Note that the samples for Gamma and Sigma^2 are on the orthonormal scale.
}
\examples{
library(fda)
x <- onechild$day
y <- onechild$height
fit <- rjmonopoly(x, y)
}
