% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plotDegreePost}
\alias{plotDegreePost}
\title{Title}
\usage{
plotDegreePost(obj, ...)
}
\arguments{
\item{obj}{object of type "rjmonopolyfit" for plotting}

\item{...}{other optional parameters for controlling plot style}
}
\value{
obj
}
\description{
Title
}
