% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin.R
\name{thin}
\alias{thin}
\title{Thin an rjmonopolyfit}
\usage{
thin(obj, thin_period, ...)
}
\arguments{
\item{obj}{object of type "rjmonopolyfit"}

\item{thin_period}{period of thinned samples}

\item{...}{other}
}
\value{
an object of type rjmonopolyfit, with n_accepted changed to
n_accepted_orig(inally)
}
\description{
Thins an object of type "rjmonopolyfit" by *thin_period*. All quantites are
thinned by the same period.
}
